fix: quiz connection mixin
